Move query to seperate file
