+++ src/pages/team.js
@@ -11,254 +11,10 @@ import H1 from "../styles/H1"
 import Img from "gatsby-image";
 import Portfolio from "../components/PortfolioSection"
 import "../styles/Global.css";
+import Query from "../GraphQLQueries/TeamPageQuery"
 
 const Team = () => {
-  const data = useStaticQuery(graphql`
-    query Team {
-      headerBackground: file(relativePath: {eq: "teamPicture.png"}) {
-        id
-        childImageSharp {
-          fluid(quality:100) {
-            ...GatsbyImageSharpFluid
-          }
-        }
-      }
-      profile: file(relativePath: {eq: "profilePicImges/AmandaAdam.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      futureview: file(relativePath: {eq: "marbles/Future View.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150, height: 150, webpQuality: 100) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      projects: file(relativePath: {eq: "marbles/Projects.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150, height: 150, webpQuality: 100) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      community: file(relativePath: {eq: "marbles/Community.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150, height: 150, webpQuality: 100) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      design: file(relativePath: {eq: "marbles/Design.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150, height: 150, webpQuality: 100) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      marketing: file(relativePath: {eq: "marbles/Marketing.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150, height: 150, webpQuality: 100) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      education: file(relativePath: {eq: "marbles/Education.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150, height: 150, webpQuality: 100) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      finance: file(relativePath: {eq: "marbles/Finance.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150, height: 150, webpQuality: 100) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      industry: file(relativePath: {eq: "marbles/Industry.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150, webpQuality: 100) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      operations: file(relativePath: {eq: "marbles/Operations.png"}) {
-        id
-        childImageSharp {
-          fixed(quality:100, width: 150, height: 150, webpQuality: 100) {
-            ...GatsbyImageSharpFixed
-          }
-        }
-      }
-      OperationsTeam: allFile(
-        filter: {
-          extension: { regex: "/(jpg)|(png)|(jpeg)/" }
-          relativeDirectory: { eq: "OperationsTeam" }
-        }
-      ) {
-        edges {
-          node {
-            base
-            childImageSharp {
-              fixed(quality:100, width: 100, height: 100, webpQuality: 100) {
-                ...GatsbyImageSharpFixed
-              }
-            }
-          }
-        }
-      }
-      MarketingTeam: allFile(
-        filter: {
-          extension: { regex: "/(jpg)|(png)|(jpeg)/" }
-          relativeDirectory: { eq: "MarketingTeam" }
-        }
-      ) {
-        edges {
-          node {
-            base
-            childImageSharp {
-              fixed(quality:100, width: 100, height: 100, webpQuality: 100) {
-                ...GatsbyImageSharpFixed
-              }
-            }
-          }
-        }
-      }
-      IndustryTeam: allFile(
-        filter: {
-          extension: { regex: "/(jpg)|(png)|(jpeg)/" }
-          relativeDirectory: { eq: "IndustryTeam" }
-        }
-      ) {
-        edges {
-          node {
-            base
-            childImageSharp {
-              fixed(quality:100, width: 100, height: 100, webpQuality: 100) {
-                ...GatsbyImageSharpFixed
-              }
-            }
-          }
-        }
-      }
-      FutureViewTeam: allFile(
-        filter: {
-          extension: { regex: "/(jpg)|(png)|(jpeg)/" }
-          relativeDirectory: { eq: "FutureViewTeam" }
-        }
-      ) {
-        edges {
-          node {
-            base
-            childImageSharp {
-              fixed(quality:100, width: 100, height: 100, webpQuality: 100) {
-                ...GatsbyImageSharpFixed
-              }
-            }
-          }
-        }
-      }
-      FinanceTeam: allFile(
-        filter: {
-          extension: { regex: "/(jpg)|(png)|(jpeg)/" }
-          relativeDirectory: { eq: "FinanceTeam" }
-        }
-      ) {
-        edges {
-          node {
-            base
-            childImageSharp {
-              fixed(quality:100, width: 100, height: 100, webpQuality: 100) {
-                ...GatsbyImageSharpFixed
-              }
-            }
-          }
-        }
-      }
-      EducationTeam: allFile(
-        filter: {
-          extension: { regex: "/(jpg)|(png)|(jpeg)/" }
-          relativeDirectory: { eq: "EducationTeam" }
-        }
-      ) {
-        edges {
-          node {
-            base
-            childImageSharp {
-              fixed(quality:100, width: 100, height: 100, webpQuality: 100) {
-                ...GatsbyImageSharpFixed
-              }
-            }
-          }
-        }
-      }
-      ProjectsTeam: allFile(
-        filter: {
-          extension: { regex: "/(jpg)|(png)|(jpeg)/" }
-          relativeDirectory: { eq: "ProjectsTeam" }
-        }
-      ) {
-        edges {
-          node {
-            base
-            childImageSharp {
-              fixed(quality:100, width: 100, height: 100, webpQuality: 100) {
-                ...GatsbyImageSharpFixed
-              }
-            }
-          }
-        }
-      }
-      DesignTeam: allFile(
-        filter: {
-          extension: { regex: "/(jpg)|(png)|(jpeg)/" }
-          relativeDirectory: { eq: "DesignTeam" }
-        }
-      ) {
-        edges {
-          node {
-            base
-            childImageSharp {
-              fixed(quality:100, width: 100, height: 100, webpQuality: 100) {
-                ...GatsbyImageSharpFixed
-              }
-            }
-          }
-        }
-      }
-      CommunityTeam: allFile(
-        filter: {
-          extension: { regex: "/(jpg)|(png)|(jpeg)/" }
-          relativeDirectory: { eq: "CommunityTeam" }
-        }
-      ) {
-        edges {
-          node {
-            base
-            childImageSharp {
-              fixed(quality:100, width: 100, height: 100, webpQuality: 100) {
-                ...GatsbyImageSharpFixed
-              }
-            }
-          }
-        }
-      }
-    }
-  `)
-
+  const data = Query()
   console.log(data)
 
   return (
